fix: toast not appearing when toast.* called from useEffect hook
## Issue
After calling a toast.* function from within a useEffect hook, the toast
does not appear (see the new test case in toast.test.tsx for an
example).

## Cause
The `useStore` hook (used by `Toaster`) calls "useState" to get the
current state of the global `memoryState`, and then calls "useEffect" to
set up a subscription to the global `memoryState`.

The bug is caused when a change to the global `memoryState` happens
after "useState" is called, but before the "useEffect" callback is run.
This can happen e.g. if we call a toast.* function from within another
component's useEffect hook.

## Solution
To setup the global `memoryState` subscription, we use the
"useLayoutEffect" hook instead. This hook runs the effect synchronously,
so no change to the global `memoryState` can happen before the effect is
run.

## Alternatives
`useSyncExternalStore` is a new hook that is designed to solve this
problem and is built into React; we could use this hook instead:
https://reactjs.org/docs/hooks-reference.html#usesyncexternalstore