fix: useSyncExternalStore instead of useState / useEffect
-- Issue
Calling a toast.* function before rendering the `<Toaster>` component,
then rendering the `<Toaster>` component, would cause the toast to not
appear or to appear "stacked up" on top of each other.

-- Cause
The `useStore` hook was using `useState` and `useEffect` to subscribe
to the global `memoryState`; unfortunately, in various cases it was
possible for change to occur to `memoryState` before the `useEffect`
callback was called. Since the `useEffect` callback hadn't run, those
changes didn't cause a re-render of the `<Toaster>` component.

-- Solution
Use the `useSyncExternalStore` hook instead of `useState` and
`useEffect`; it handles these cases correctly.

I've added a dependency on the `use-sync-external-store` shim so that
React version below 18 are still supported.

-- Related
https://github.com/timolins/react-hot-toast/issues/253